% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jian_models.R
\name{download_jian_fits}
\alias{download_jian_fits}
\title{Download full bootstrap (500 fits) of model fits from the Open
Science Framework (OSF)}
\usage{
download_jian_fits(destfile, ...)
}
\arguments{
\item{destfile}{a character string (or vector, see \code{url}) with
    the name where the downloaded file is saved.  Tilde-expansion is
    performed.}

\item{...}{Additional arguments to \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
\code{destfile}, invisibly
}
\description{
The project is hosted at https://osf.io/fzrcq/.
}
