% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jian_models.R
\name{fit_jian_ann}
\alias{fit_jian_ann}
\alias{fit_jian_rf}
\title{Fit models from Jian et al. paper}
\usage{
fit_jian_ann(data, use_rock = FALSE, verbose = FALSE)

fit_jian_rf(data, use_rock = FALSE, top_type = FALSE)
}
\arguments{
\item{data}{Input data for fit.}

\item{use_rock}{Logical. If \code{TRUE}, include the \code{Percent_Rock} as a predictor.}

\item{verbose}{Logical. If \code{TRUE}, print the number of attempts.}

\item{top_type}{Logical. If \code{TRUE}, include the \code{Top_Type} as a
predictor in the Random Forest model.}
}
\value{
For \code{fit_jian_ann}, the output of \code{\link[neuralnet:neuralnet]{neuralnet::neuralnet()}},
but with the special class attribute \code{urbankfs_ann} (so we can
define our own \code{predict} S3 method). For \code{fit_jian_rf}, the
output of \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.
}
\description{
\code{fit_jian_ann} fits an artificial neural network, and \code{fit_jian_rf}
fits a random forest model (with or without \code{top_type} included).
}
\author{
Alexey Shiklomanov
}
