% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_soil_pct.R
\name{normalize_soil_pct}
\alias{normalize_soil_pct}
\alias{normalize_soil_pct_data}
\title{Normalize soil sand, silt, clay, and rock percentages to 100%}
\usage{
normalize_soil_pct(Percent_Sand, Percent_Silt, Percent_Clay,
  Percent_Rock_Fragment = NULL)

normalize_soil_pct_data(df, add_rock = TRUE)
}
\arguments{
\item{Percent_Sand, Percent_Silt, Percent_Clay}{Percentages (0 - 100%) of sand, silt, and clay in soil}

\item{Percent_Rock_Fragment}{Percent rock in soil. If \code{NULL} (default), re-normalize sand,
silt, and clay to 100%}

\item{df}{\code{data.frame} containing soil percentages in columns
\code{soil_cols}}

\item{add_rock}{(Logical) If \code{TRUE}, add rock data}
}
\description{
If no \code{Percent_Rock_Fragment} is given (\code{Percent_Rock_Fragment = NULL}, default), renormalize sand, silt, and
clay percents so that they sum to 100%. If rock is given, normalize sand,
silt, and clay so that their sum = \code{Percent_Rock_Fragment - 100}.
}
\author{
Alexey Shiklomanov
}
