% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bootstrap.R
\name{predict_bootstrap}
\alias{predict_bootstrap}
\title{Predict soil conductivity for a set of bootstrapped models}
\usage{
predict_bootstrap(data, fitted_models = fitted_models)
}
\arguments{
\item{data}{Soil structure data for prediction. Must be a
\code{data.frame} with columns \code{Percent_Sand}, \code{Percent_Silt},
\code{Percent_Clay}, and \code{Top_Type}.}

\item{fitted_models}{Nested \code{data.frame} containing fitted models.
Default is package data.}
}
\value{
Nested \code{data.frame} of predicted soil conductivities for
each model sample.
}
\description{
Predict soil conductivity for a set of bootstrapped models
}
\examples{
df <- data.frame(
  Percent_Sand = c(14, 15, 18, 18),
  Percent_Silt = c(63, 15, 59, 60),
  Percent_Clay = c(23, 70, 23, 22),
  Top_Type = rep("blocky", 4),
  stringsAsFactors = FALSE
)
raw_pred <- predict_bootstrap(df)
# Tidy summary
summary(raw_pred)
}
\author{
Alexey Shiklomanov
}
