% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for ANN or random forest models}
\usage{
prepare_data(data, use_rock = FALSE, top_type = FALSE)
}
\arguments{
\item{data}{Input data for fit.}

\item{use_rock}{Logical. If \code{TRUE}, include the \code{Percent_Rock_Fragment} as a predictor.}

\item{top_type}{Logical. If \code{TRUE}, include the \code{Top_Type} as a
predictor in the Random Forest model.}
}
\value{
\code{data.frame} with only the necessary columns and no missing data
}
\description{
Prepare data for ANN or random forest models
}
\author{
Alexey Shiklomanov
}
