% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jian_models.R
\name{pretty_model_types}
\alias{pretty_model_types}
\title{Nicer factor levels for model type}
\usage{
pretty_model_types(pretty = c("name", "value"))
}
\arguments{
\item{pretty}{Where should the pretty name be? If \code{"name"}
(default), the pretty label is the name and value is the original
(useful for, e.g., \code{\link[forcats:fct_recode]{forcats::fct_recode()}}). If \code{"value"}, the
reverse.}
}
\value{
Character vector with
}
\description{
Nicer factor levels for model type
}
\author{
Alexey Shiklomanov
}
