% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pvalue}
\alias{pvalue}
\title{Extract p value from model object}
\usage{
pvalue(object)
}
\arguments{
\item{object}{Model fit object (e.g. output of \code{\link[stats:lm]{stats::lm()}})}
}
\value{
Model p value, based on f-statistic
}
\description{
Extract p value from model object
}
