% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_length}
\alias{same_length}
\title{Check that objects have the same length}
\usage{
same_length(...)
}
\arguments{
\item{...}{Objects}
}
\value{
\code{TRUE} is all elements in \code{...} have the same length, and \code{FALSE}
otherwise
}
\description{
Check that objects have the same length
}
\author{
Alexey Shiklomanov
}
