% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_range.R
\name{scale_range}
\alias{scale_range}
\alias{unscale_range}
\title{Re-scale and un-scale data based on its range}
\usage{
scale_range(x)

unscale_range(x, y = x)
}
\arguments{
\item{x}{Vector to scale or un-scale}

\item{y}{Vector whose limits will be used for re-scaling}
}
\value{
Object of same type as \code{x}
}
\description{
Re-scale and un-scale data based on its range
}
\author{
Alexey Shiklomanov
}
