% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shrink_randomforest}
\alias{shrink_randomforest}
\title{Shrink the size of a \code{randomForest} model object by removing pieces
not necessary for prediction}
\usage{
shrink_randomforest(object)
}
\arguments{
\item{object}{Fitted \code{\link[randomForest:randomForest]{randomForest::randomForest()}} model}
}
\value{
\code{randomForest} object with excess data removed
}
\description{
The large pieces are \code{proximity} and \code{terms}. The \code{predict} method
for \code{randomForest} objects should be robust to these being \code{NULL}.
}
