% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_types.R
\name{soil_type_levels}
\alias{soil_type_levels}
\title{Alphabetical list of soil types, to be used as factor levels}
\usage{
soil_type_levels()
}
\value{
Character vector of soil types, in alphabetical order
}
\description{
Alphabetical list of soil types, to be used as factor levels
}
