% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_types.R
\name{soil_types}
\alias{soil_types}
\title{Conversion table between specific and general soil types}
\usage{
soil_types(col = NULL, stringsAsFactors = FALSE)
}
\arguments{
\item{col}{Name of column to return. If \code{NULL} (default), return
the full \code{data.frame}}

\item{stringsAsFactors}{logical: should character vectors be converted
    to factors?  The \sQuote{factory-fresh} default is \code{TRUE}, but
    this can be changed by setting \code{\link{options}(stringsAsFactors
      = FALSE)}.}
}
\value{
If \code{is.null(col)}, \code{data.frame} linking specific (\code{Type})
and general (\code{Top_Type}) soils
}
\description{
Conversion table between specific and general soil types
}
\author{
Alexey Shiklomanov
}
