% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bootstrap.R
\name{summary.urbankfs_prediction}
\alias{summary.urbankfs_prediction}
\title{Summary method for output of \code{\link[=predict_bootstrap]{predict_bootstrap()}}.}
\usage{
\method{summary}{urbankfs_prediction}(object, quantiles = c(0.05, 0.5,
  0.95), ...)
}
\arguments{
\item{object}{Output of \code{\link[=predict_bootstrap]{predict_bootstrap()}}}

\item{quantiles}{Numeric vector of quantiles for summary
statistics. Default = \code{c(0.05, 0.5, 0.95)}.}
}
\value{
\code{data.frame} of prediction summaries
}
\description{
Summary method for output of \code{\link[=predict_bootstrap]{predict_bootstrap()}}.
}
