% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_soil_pct.R
\name{normalize_soil_pct}
\alias{normalize_soil_pct}
\alias{normalize_soil_pct_data}
\title{Normalize soil sand, silt, and clay percentages to 100%}
\usage{
normalize_soil_pct(sand, silt, clay)

normalize_soil_pct_data(df, soil_cols = paste0("Percent_", c("Sand",
  "Silt", "Clay")))
}
\arguments{
\item{sand, silt, clay}{Percentages (0 - 100%) of sand, silt, and clay in soil}

\item{df}{\code{data.frame} containing soil percentages in columns
\code{soil_cols}}

\item{soil_cols}{Column names corresponding to sand, silt, and clay}
}
\description{
Normalize soil sand, silt, and clay percentages to 100%
}
\author{
Alexey Shiklomanov
}
